# Optimize W256 serialization